refactor(uuid-hash): moved hash format to utils file
