+++ lib/core/shards/MongoDBShardsRepository.ts
@@ -1,4 +1,3 @@
-import { validate } from 'uuid';
 import { Contact } from '../contacts/Contact';
 import { ShardsRepository } from './Repository';
 import { Shard } from './Shard';
@@ -14,8 +13,6 @@ const formatFromMongoToShard = (mongoShard: any): Shard => {
       nodeID: c.nodeID,
     });
   });
-
-  
   const formattedShard = {
     ...shard,
     contracts,
@@ -34,27 +31,6 @@ const formatFromMongoToShard = (mongoShard: any): Shard => {
 
   return formattedShard;
 };
-
-
-export const formatShardHash = (shard: Shard): Shard => {
-  if (shard.hash) {
-    return { ...shard, hash: removeUuidFromHash(shard.hash) };
-  }
-  return shard;
-}
-
-export const removeUuidFromHash = (hash: string): string => {
-  const separatorIndex = hash.indexOf('$');
-  if (separatorIndex !== -1) {
-    const potentialUuid = hash.slice(0, separatorIndex);
-
-    if (validate(potentialUuid)) {
-      return hash.slice(separatorIndex + 1)
-    }
-  }
-  return hash;
-}
-
 export class MongoDBShardsRepository implements ShardsRepository {
   constructor(private model: any) {}
