+++ lib/server/routes/buckets.js
@@ -27,7 +27,8 @@ const { BucketEntriesUsecase, BucketEntryVersionNotFoundError } = require('../..
 const { MongoDBBucketEntryShardsRepository } = require('../../core/bucketEntryShards/MongoDBBucketEntryShardsRepository');
 const { MongoDBMirrorsRepository } = require('../../core/mirrors/MongoDBMirrorsRepository');
 const { MongoDBFramesRepository } = require('../../core/frames/MongoDBFramesRepository');
-const { MongoDBShardsRepository, formatShardHash } = require('../../core/shards/MongoDBShardsRepository');
+const { MongoDBShardsRepository } = require('../../core/shards/MongoDBShardsRepository');
+const { formatShardHash } = require('../../core/shards/utils');
 const { MongoDBBucketsRepository, MongoDBUsersRepository } = require('../../core');
 const { MongoDBUploadsRepository } = require('../../core/uploads/MongoDBUploadsRepository');
 const { MongoDBPointersRepository } = require('../../core/pointers/MongoDBPointersRepository');
